Added some more nsfdata structs and fixed a number of bugs

+++ org.openntf.domino/src/org/openntf/domino/nsfdata/structs/FRAMESETLENGTH.java
@@ -44,7 +44,7 @@ public class FRAMESETLENGTH extends AbstractStruct {
 	}
 
 	@Override
-	public int getStructSize() {
+	public long getStructSize() {
 		return SIZE;
 	}
