+++ org.openntf.domino/src/org/openntf/domino/nsfdata/structs/OLE_GUID.java
@@ -25,7 +25,7 @@ public class OLE_GUID extends AbstractStruct {
 	}
 
 	@Override
-	public int getStructSize() {
+	public long getStructSize() {
 		return SIZE;
 	}
