+++ org.openntf.domino/src/org/openntf/domino/nsfdata/structs/RECTSIZE.java
@@ -27,7 +27,7 @@ public class RECTSIZE extends AbstractStruct {
 	}
 
 	@Override
-	public int getStructSize() {
+	public long getStructSize() {
 		return SIZE;
 	}
