+++ org.openntf.domino/src/org/openntf/domino/nsfdata/structs/TFMT.java
@@ -22,7 +22,7 @@ public class TFMT extends AbstractStruct {
 	}
 
 	@Override
-	public int getStructSize() {
+	public long getStructSize() {
 		return SIZE;
 	}
