+++ org.openntf.domino/src/org/openntf/domino/nsfdata/structs/cd/CDFONTTABLE.java
@@ -63,7 +63,7 @@ public class CDFONTTABLE extends CDRecord {
 		}
 
 		@Override
-		public int getStructSize() {
+		public long getStructSize() {
 			return SIZE;
 		}
