+++ org.openntf.domino/src/org/openntf/domino/nsfdata/structs/cd/CDTEXT.java
@@ -30,7 +30,7 @@ public class CDTEXT extends CDRecord {
 	}
 
 	public int getTextLength() {
-		return getDataLength() - 4;
+		return (int) (getDataLength() - 4);
 	}
 
 	@Override
